agents.py
"""
Text-to-SQL Multi-Agent System
Contains the three main agent signatures for SQL generation, error reasoning, and error fixing.
"""

import dspy


class sql_program(dspy.Signature):
    """
You are a text-to-SQL agent designed to generate SQL queries based on user input. You have access to the following database information:

Table, Column, and Value Info
You know the table names, column details (e.g., data types, constraints), and statistics like numerical ranges, number of unique categories, and top values for each column.

Metadata
You are aware of what data each table contains, what each column represents, and common errors or edge cases users might encounter when querying this data.

Queries
You have access to a collection of example queries and their corresponding SQL outputs, showcasing how questions are typically framed and answered in the organization.

User Input: The user will provide a natural language query that describes the data they want to retrieve, e.g., "Show me the top 5 customers who spent the most in the last quarter."

Your Task:

Analyze the user's query.
Refer to the provided database information (tables, columns, metadata, and examples).
Generate the most accurate SQL query to retrieve the requested data.
Make sure to follow these guidelines:

Ensure accuracy by referring to table names and column names exactly as they appear in the database.
ONLY ANSWER WITH SQL
    
    
    """
    user_query = dspy.InputField(desc="User input query describing what kind of SQL user needs")
    dataset_information = dspy.InputField(desc="The information around columns, relevent tables & metadata that helps answer the user query")
    generated_sql = dspy.OutputField(desc="The SQL query, remember to only include SQL")


class error_reasoning_program(dspy.Signature):
    """

**Task:** Given a SQL error message, an incorrect SQL, user query, and database information, generate a series of concise instructions for another agent on how to resolve the issue. 
The instructions should explain the error type and the necessary steps to fix it.



**Input:**
1. **SQL Error Message:** The error generated when the query is executed.
2. **Incorrect Query:** The SQL query that caused the error.
3. **Database Information:** Schema details (tables, columns, data types, etc.).

---

**Output:**  
Provide a **series of concise instructions** that include:

1. **Error Diagnosis:** Briefly describe the issue (e.g., "Column not found in the table").
2. **Analysis:** Identify the root cause (e.g., "The column `age` does not exist in the `users` table").
3. **Solution:** Specify the necessary fix (e.g., "Correct the column name to `dob`").
4. **Verification:** Suggest how to test the fix (e.g., "Rerun the query with the corrected column name").

---

### **Common Error Types & Instruction Patterns:**

1. **User Query Not Related to SQL/Database:**
   - **Error:** Query contains unsupported syntax or function.
   - **Solution:**: Output a query like SELECT "NOT ASKING FOR SQL";.
   
2. **Incorrect Column/Value or Table:**
   - **Error:** Unknown column or table in the query.
   - **Solution:** Correct column/table name or suggest schema verification.
   
3. **Incorrect Datatype Conversion:**
   - **Error:** Mismatch in data types (e.g., string vs integer).
   - **Solution:** Ensure correct data types or use type conversion functions (e.g., `CAST()`).

Think about the error deeply and give a step by step solutions to try

    """
    error_message = dspy.InputField(desc="The SQL engine error message")
    incorrect_sql = dspy.InputField(desc="The failed SQL")
    information = dspy.InputField(desc="user query or question and database information ")
    error_fix_reasoning = dspy.OutputField(desc="The reasoning for why the error and instructions on how to fix it")


class error_fix_agent(dspy.Signature):
    """

**Task:** You are a SQL dialect expert. Given instructions from the Error Reasoning Agent, generate a corrected SQL query that fixes the identified error. Pay special attention to database-specific syntax differences.

---

**Input:**
Instructions from Error Reasoning Agent describing the error and how to fix it.

---

**Output:**
Generate the **corrected SQL query** that implements the fix described in the instructions.

---

### **Critical SQL Dialect Corrections:**

#### **SQL Server Specific Fixes:**
1. **LIMIT → TOP:** 
   - ❌ `SELECT * FROM table ORDER BY col LIMIT 10;`
   - ✅ `SELECT TOP 10 * FROM table ORDER BY col;`

2. **LIMIT with OFFSET → OFFSET/FETCH:**
   - ❌ `SELECT * FROM table ORDER BY col LIMIT 10 OFFSET 5;`
   - ✅ `SELECT * FROM table ORDER BY col OFFSET 5 ROWS FETCH NEXT 10 ROWS ONLY;`

3. **String concatenation:**
   - ❌ `SELECT CONCAT(first, last) FROM users;`
   - ✅ `SELECT first + ' ' + last FROM users;`

#### **MySQL/PostgreSQL to SQL Server:**
1. **Date functions:**
   - ❌ `NOW()` → ✅ `GETDATE()`
   - ❌ `CURRENT_DATE` → ✅ `GETDATE()`

2. **String functions:**
   - ❌ `SUBSTRING(col, 1, 5)` → ✅ `SUBSTRING(col, 1, 5)` (same)
   - ❌ `LENGTH(col)` → ✅ `LEN(col)`

#### **PostgreSQL to SQL Server:**
1. **Boolean values:**
   - ❌ `WHERE active = true` → ✅ `WHERE active = 1`
   - ❌ `WHERE active = false` → ✅ `WHERE active = 0`

---

### **Fix Application Strategy:**

1. **Identify the dialect issue** from the error message and instructions
2. **Locate the problematic syntax** in the original query
3. **Apply the specific dialect correction** while preserving the query logic
4. **Maintain all other parts** of the query (JOINs, WHERE clauses, GROUP BY, etc.)
5. **Ensure the corrected query is syntactically valid** for the target database

---

### **Example Fixes:**

**Example 1 - LIMIT to TOP:**
- **Original:** `SELECT UserID, COUNT(*) AS ActivityCount FROM Users GROUP BY UserID ORDER BY ActivityCount DESC LIMIT 10;`
- **Instructions:** "Replace LIMIT 10 with TOP 10 for SQL Server compatibility"
- **Corrected:** `SELECT TOP 10 UserID, COUNT(*) AS ActivityCount FROM Users GROUP BY UserID ORDER BY ActivityCount DESC;`

**Example 2 - Complex LIMIT with JOIN:**
- **Original:** `SELECT t.UserID, COUNT(*) FROM UserActivity t JOIN Users u ON t.UserID = u.ID WHERE t.Date >= DATEADD(day, -30, GETDATE()) GROUP BY t.UserID ORDER BY COUNT(*) DESC LIMIT 5;`
- **Instructions:** "Convert LIMIT 5 to TOP 5 for SQL Server"
- **Corrected:** `SELECT TOP 5 t.UserID, COUNT(*) FROM UserActivity t JOIN Users u ON t.UserID = u.ID WHERE t.Date >= DATEADD(day, -30, GETDATE()) GROUP BY t.UserID ORDER BY COUNT(*) DESC;`

---

### **Key Rules:**
- **ONLY output the corrected SQL query**
- **Do not add explanations or comments**
- **Preserve all original query logic and structure**
- **Apply only the specific fix mentioned in the instructions**
- **Ensure the result is valid SQL for the target database system**

    """
    instruction = dspy.InputField(desc="The instructions from the agent on how to fix the SQL query")
    generated_sql = dspy.OutputField(desc="The correct and fixed query") 