requirements.txt
dspy-ai>=2.4.0
pandas>=1.5.0
duckdb>=0.9.0
openai>=1.0.0
python-dotenv>=1.0.0
jupyter>=1.0.0
tabulate>=0.9.0

# Web UI dependencies
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Database drivers
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
pyodbc>=4.0.0
pymssql>=2.2.0
sqlalchemy>=2.0.0

# Additional utilities
cryptography>=3.4.0
requests>=2.28.0 